client/game/: Moved all the game logic to it's own package
So it makes development easy and it's also a nice separation of concepts
