+++ action/type_string.go
@@ -8,11 +8,11 @@ import (
 	"strings"
 )
 
-const _TypeName = "cursor_movecamera_zoomsummon_unittpsremove_unitsteal_liveplace_towerremove_towerselect_towerselected_towerselected_tower_invaliddeselect_towerincome_ticktower_attackunit_killedwindow_resizingnavigate_tostart_gameopen_tower_menuclose_tower_menugo_homechecked_pathchange_unit_linesign_up_erroruser_sign_upuser_sign_inuser_sign_outjoin_waiting_roomexit_waiting_roomtoggle_statsadd_playerremove_playersync_statesync_userswait_room_countdown_ticksync_waiting_room"
+const _TypeName = "cursor_movecamera_zoomsummon_unittpsremove_unitsteal_liveplace_towerremove_towerselect_towerselected_towerselected_tower_invaliddeselect_towerincome_ticktower_attackunit_killedwindow_resizingnavigate_tostart_gameopen_tower_menuclose_tower_menugo_homechange_unit_linesign_up_erroruser_sign_upuser_sign_inuser_sign_outjoin_waiting_roomexit_waiting_roomtoggle_statsadd_playerremove_playersync_statesync_userswait_room_countdown_ticksync_waiting_room"
 
-var _TypeIndex = [...]uint16{0, 11, 22, 33, 36, 47, 57, 68, 80, 92, 106, 128, 142, 153, 165, 176, 191, 202, 212, 227, 243, 250, 262, 278, 291, 303, 315, 328, 345, 362, 374, 384, 397, 407, 417, 441, 458}
+var _TypeIndex = [...]uint16{0, 11, 22, 33, 36, 47, 57, 68, 80, 92, 106, 128, 142, 153, 165, 176, 191, 202, 212, 227, 243, 250, 266, 279, 291, 303, 316, 333, 350, 362, 372, 385, 395, 405, 429, 446}
 
-const _TypeLowerName = "cursor_movecamera_zoomsummon_unittpsremove_unitsteal_liveplace_towerremove_towerselect_towerselected_towerselected_tower_invaliddeselect_towerincome_ticktower_attackunit_killedwindow_resizingnavigate_tostart_gameopen_tower_menuclose_tower_menugo_homechecked_pathchange_unit_linesign_up_erroruser_sign_upuser_sign_inuser_sign_outjoin_waiting_roomexit_waiting_roomtoggle_statsadd_playerremove_playersync_statesync_userswait_room_countdown_ticksync_waiting_room"
+const _TypeLowerName = "cursor_movecamera_zoomsummon_unittpsremove_unitsteal_liveplace_towerremove_towerselect_towerselected_towerselected_tower_invaliddeselect_towerincome_ticktower_attackunit_killedwindow_resizingnavigate_tostart_gameopen_tower_menuclose_tower_menugo_homechange_unit_linesign_up_erroruser_sign_upuser_sign_inuser_sign_outjoin_waiting_roomexit_waiting_roomtoggle_statsadd_playerremove_playersync_statesync_userswait_room_countdown_ticksync_waiting_room"
 
 func (i Type) String() string {
 	if i < 0 || i >= Type(len(_TypeIndex)-1) {
@@ -46,24 +46,23 @@ func _TypeNoOp() {
 	_ = x[OpenTowerMenu-(18)]
 	_ = x[CloseTowerMenu-(19)]
 	_ = x[GoHome-(20)]
-	_ = x[CheckedPath-(21)]
-	_ = x[ChangeUnitLine-(22)]
-	_ = x[SignUpError-(23)]
-	_ = x[UserSignUp-(24)]
-	_ = x[UserSignIn-(25)]
-	_ = x[UserSignOut-(26)]
-	_ = x[JoinWaitingRoom-(27)]
-	_ = x[ExitWaitingRoom-(28)]
-	_ = x[ToggleStats-(29)]
-	_ = x[AddPlayer-(30)]
-	_ = x[RemovePlayer-(31)]
-	_ = x[SyncState-(32)]
-	_ = x[SyncUsers-(33)]
-	_ = x[WaitRoomCountdownTick-(34)]
-	_ = x[SyncWaitingRoom-(35)]
+	_ = x[ChangeUnitLine-(21)]
+	_ = x[SignUpError-(22)]
+	_ = x[UserSignUp-(23)]
+	_ = x[UserSignIn-(24)]
+	_ = x[UserSignOut-(25)]
+	_ = x[JoinWaitingRoom-(26)]
+	_ = x[ExitWaitingRoom-(27)]
+	_ = x[ToggleStats-(28)]
+	_ = x[AddPlayer-(29)]
+	_ = x[RemovePlayer-(30)]
+	_ = x[SyncState-(31)]
+	_ = x[SyncUsers-(32)]
+	_ = x[WaitRoomCountdownTick-(33)]
+	_ = x[SyncWaitingRoom-(34)]
 }
 
-var _TypeValues = []Type{CursorMove, CameraZoom, SummonUnit, TPS, RemoveUnit, StealLive, PlaceTower, RemoveTower, SelectTower, SelectedTower, SelectedTowerInvalid, DeselectTower, IncomeTick, TowerAttack, UnitKilled, WindowResizing, NavigateTo, StartGame, OpenTowerMenu, CloseTowerMenu, GoHome, CheckedPath, ChangeUnitLine, SignUpError, UserSignUp, UserSignIn, UserSignOut, JoinWaitingRoom, ExitWaitingRoom, ToggleStats, AddPlayer, RemovePlayer, SyncState, SyncUsers, WaitRoomCountdownTick, SyncWaitingRoom}
+var _TypeValues = []Type{CursorMove, CameraZoom, SummonUnit, TPS, RemoveUnit, StealLive, PlaceTower, RemoveTower, SelectTower, SelectedTower, SelectedTowerInvalid, DeselectTower, IncomeTick, TowerAttack, UnitKilled, WindowResizing, NavigateTo, StartGame, OpenTowerMenu, CloseTowerMenu, GoHome, ChangeUnitLine, SignUpError, UserSignUp, UserSignIn, UserSignOut, JoinWaitingRoom, ExitWaitingRoom, ToggleStats, AddPlayer, RemovePlayer, SyncState, SyncUsers, WaitRoomCountdownTick, SyncWaitingRoom}
 
 var _TypeNameToValueMap = map[string]Type{
 	_TypeName[0:11]:         CursorMove,
@@ -108,36 +107,34 @@ var _TypeNameToValueMap = map[string]Type{
 	_TypeLowerName[227:243]: CloseTowerMenu,
 	_TypeName[243:250]:      GoHome,
 	_TypeLowerName[243:250]: GoHome,
-	_TypeName[250:262]:      CheckedPath,
-	_TypeLowerName[250:262]: CheckedPath,
-	_TypeName[262:278]:      ChangeUnitLine,
-	_TypeLowerName[262:278]: ChangeUnitLine,
-	_TypeName[278:291]:      SignUpError,
-	_TypeLowerName[278:291]: SignUpError,
-	_TypeName[291:303]:      UserSignUp,
-	_TypeLowerName[291:303]: UserSignUp,
-	_TypeName[303:315]:      UserSignIn,
-	_TypeLowerName[303:315]: UserSignIn,
-	_TypeName[315:328]:      UserSignOut,
-	_TypeLowerName[315:328]: UserSignOut,
-	_TypeName[328:345]:      JoinWaitingRoom,
-	_TypeLowerName[328:345]: JoinWaitingRoom,
-	_TypeName[345:362]:      ExitWaitingRoom,
-	_TypeLowerName[345:362]: ExitWaitingRoom,
-	_TypeName[362:374]:      ToggleStats,
-	_TypeLowerName[362:374]: ToggleStats,
-	_TypeName[374:384]:      AddPlayer,
-	_TypeLowerName[374:384]: AddPlayer,
-	_TypeName[384:397]:      RemovePlayer,
-	_TypeLowerName[384:397]: RemovePlayer,
-	_TypeName[397:407]:      SyncState,
-	_TypeLowerName[397:407]: SyncState,
-	_TypeName[407:417]:      SyncUsers,
-	_TypeLowerName[407:417]: SyncUsers,
-	_TypeName[417:441]:      WaitRoomCountdownTick,
-	_TypeLowerName[417:441]: WaitRoomCountdownTick,
-	_TypeName[441:458]:      SyncWaitingRoom,
-	_TypeLowerName[441:458]: SyncWaitingRoom,
+	_TypeName[250:266]:      ChangeUnitLine,
+	_TypeLowerName[250:266]: ChangeUnitLine,
+	_TypeName[266:279]:      SignUpError,
+	_TypeLowerName[266:279]: SignUpError,
+	_TypeName[279:291]:      UserSignUp,
+	_TypeLowerName[279:291]: UserSignUp,
+	_TypeName[291:303]:      UserSignIn,
+	_TypeLowerName[291:303]: UserSignIn,
+	_TypeName[303:316]:      UserSignOut,
+	_TypeLowerName[303:316]: UserSignOut,
+	_TypeName[316:333]:      JoinWaitingRoom,
+	_TypeLowerName[316:333]: JoinWaitingRoom,
+	_TypeName[333:350]:      ExitWaitingRoom,
+	_TypeLowerName[333:350]: ExitWaitingRoom,
+	_TypeName[350:362]:      ToggleStats,
+	_TypeLowerName[350:362]: ToggleStats,
+	_TypeName[362:372]:      AddPlayer,
+	_TypeLowerName[362:372]: AddPlayer,
+	_TypeName[372:385]:      RemovePlayer,
+	_TypeLowerName[372:385]: RemovePlayer,
+	_TypeName[385:395]:      SyncState,
+	_TypeLowerName[385:395]: SyncState,
+	_TypeName[395:405]:      SyncUsers,
+	_TypeLowerName[395:405]: SyncUsers,
+	_TypeName[405:429]:      WaitRoomCountdownTick,
+	_TypeLowerName[405:429]: WaitRoomCountdownTick,
+	_TypeName[429:446]:      SyncWaitingRoom,
+	_TypeLowerName[429:446]: SyncWaitingRoom,
 }
 
 var _TypeNames = []string{
@@ -162,21 +159,20 @@ var _TypeNames = []string{
 	_TypeName[212:227],
 	_TypeName[227:243],
 	_TypeName[243:250],
-	_TypeName[250:262],
-	_TypeName[262:278],
-	_TypeName[278:291],
+	_TypeName[250:266],
+	_TypeName[266:279],
+	_TypeName[279:291],
 	_TypeName[291:303],
-	_TypeName[303:315],
-	_TypeName[315:328],
-	_TypeName[328:345],
-	_TypeName[345:362],
-	_TypeName[362:374],
-	_TypeName[374:384],
-	_TypeName[384:397],
-	_TypeName[397:407],
-	_TypeName[407:417],
-	_TypeName[417:441],
-	_TypeName[441:458],
+	_TypeName[303:316],
+	_TypeName[316:333],
+	_TypeName[333:350],
+	_TypeName[350:362],
+	_TypeName[362:372],
+	_TypeName[372:385],
+	_TypeName[385:395],
+	_TypeName[395:405],
+	_TypeName[405:429],
+	_TypeName[429:446],
 }
 
 // TypeString retrieves an enum value from the enum constants string name.
