--- client/colors.go
@@ -1,8 +1,0 @@
-package client
-
-import "image/color"
-
-var (
-	green = color.RGBA{0x00, 0x80, 0x00, 0xff}
-	red   = color.RGBA{0x80, 0x00, 0x00, 0xff}
-)
