+++ client/utils/colors.go
@@ -1,0 +1,8 @@
+package utils
+
+import "image/color"
+
+var (
+	Green = color.RGBA{0x00, 0x80, 0x00, 0xff}
+	Red   = color.RGBA{0x80, 0x00, 0x00, 0xff}
+)
