+++ client/utils/routes.go
@@ -1,0 +1,8 @@
+package utils
+
+const (
+	SignUpRoute      = "sign_up"
+	LobbyRoute       = "lobby"
+	GameRoute        = "game"
+	WaitingRoomRoute = "waiting_room"
+)
